+++ lib/screens/add_book_screen.dart
@@ -1,6 +1,5 @@
-import 'package:agnes_app/requests.dart';
+import 'package:agnes_app/Generic/requests.dart';
 import 'package:flutter/material.dart';
-
 
 class AddNewBook extends StatefulWidget {
   const AddNewBook({Key? key}) : super(key: key);
@@ -69,7 +68,7 @@ class _AskIsbnCodeState extends State<_AskIsbnCode> {
                     decoration: const InputDecoration(
                       icon: Icon(Icons.format_size_sharp),
                       helperText:
-                      'Você encontra o código ISBN próximo ao código de barras, no verso do livro. \nPor exemplo: 978-8576573937',
+                          'Você encontra o código ISBN próximo ao código de barras, no verso do livro. \nPor exemplo: 978-8576573937',
                       helperMaxLines: 4,
                       labelText: 'Digite o código ISBN',
                     ),
